Changed text mode paragraph indentation behavior
Up until now, I had been taking advantage of paragraph indentation in text
mode by running (paragraph-indent-minor-mode) as a text-mode hook.  This
had the negative side-effect that it would screw up auto indentation
behavior in org-mode (which is based on text-mode), and in theory it could
get in the way elsewhere as well.

I've solved this issue by removing said hook from text-mode, and instead
adding the major mode (paragraph-indent-text-mode) to the auto mode list
for files with the .txt name extension.