Increased Ido max directory size
Changed ido-max-directory-size to 100000, up from its default value of
30000.  This number determines the maximum Unix directory size, in bytes,
for which Ido will automatically enable auto-completion.  This value was
less than the size of my home directory on Noatun, so I would have to press
C-a every time I wanted to enable Ido completion in my home directory,
which was quite annoying.