Changed Slime loading to autoload
I had been loading slime with "require" in my .emacs file, but this
significantly slowed down my Emacs startup time.  This was particularly
undesirable as I do not use Slime very frequently.

I have changed the load mechanism to a test for slime with "featurep", then
an autoload command to load the module when and if it is needed.