Disable Ido Mode directory merging
Prevent Ido Mode from automatically finding a file of the same name in a
previously visited path, when the name of the file you'd like to visit does
not exist in the currently specified directory.

Before setting this option, I had the problem that if, say, I wanted to
create a file called outline.org to contain a general outline for some
project I'm working on, rather than allow me to create a new outline.org in
the current project's directory, Ido Mode would insist on opening a
different file named outline.org in one of the previous projects I had
visited.  This was rather annoying, and it made it way too easy for me to
accidentally overwrite files elsewhere in my directory tree.