Added prompt-user-selection, perl-sub-label
Added a generic prompt-user-selection function to .emacs, which provides a
shortcut for interactive functions to prompt the user for a choice from a
given set of selections.

Also added the function perl-sub-label and bound it to the keystroke \C-cj
in CPerl mode.  This function uses prompt-user-selection to generate a
standard comment block, as recommended in _Perl Best Practices_, above the
current sub definition.