Added SLIME/SWANK support for Clojure through ELPA
Installed ELPA, the Emacs Lisp package system, and used this to upgrade to
a newer version of SLIME with support for Clojure via swank-clojure.  Other
SWANK ports do not appear to be provided through ELPA, however, so I had to
copy the e.g. CLISP SWANK implementations over manually.  This results in a
warning message when starting SLIME for CLISP, but it appears to work
okay...