Quick hack to insert timestamps with TZ on Windows
For whatever reason, Emacs's format-time-string function is apparently
unable to resolve the time zone name on Windows using the "%Z" format
specifier (at least as of Emacs 23.1 on Windows 7).  This meant that
timestamps inserted by e.g. my new-journal-entry keyboard shortcut would be
missing time zone information.

This patch modifies my timestamp-string function so that if Emacs can't
figure out the time zone name on its own, it will assume a time zone offset
of "-0400" to mean EDT and "-0500" to mean EST.  A quick hack that will
only work in U.S. Eastern time, but it gets the job done for now...