Always launch new aspell process for flyspell
The flyspell-large-region variable sets a minimum threshold for starting a
new ispell/aspell process when checking a region in flyspell mode.  For
regions smaller than this threshold, flyspell uses a slower checking method
which can also detect doubled words.

The problem is that on Microsoft Windows, flyspell's small region method is
unbearably slow.  This patch sets the large region threshold to zero so
that flyspell checking is always performed in a new aspell process.