Rename init files to use default search paths
I recently learned that Emacs will automatically search for an
~/.emacs.d/init.el file if ~/.emacs or ~/.emacs.el is not present on the
system.  This commit renames ~/.emacs.d/emacs to ~/.emacs.d/init.el to
remove the need for a ~/.emacs => ~/.emacs.d/emacs symlink to be installed
on the system.

One possibility pointed out to me in a thread on the /r/emacs subreddit was
to change my machine-local init script to ~/.emacs.el, and have it call
~/.emacs.d/init.el at the end; that way I wouldn't have to explicitly load
~/.emacs.local in my version-controlled script.  However, this would mean
that the machine-local script gets loaded before any of the main script
runs, so I wouldn't be able to take advantage of custom load paths, etc. in
the local script.

However, this commit also renames ~/.emacs.local to ~/.emacs.local.el for
consistency.