Reconfigured Slime for Clojure compatibility
This commit gets rid of inferior lisp mode in favor of slime-repl for using
SLIME.  This results in duplicate buffers (a slime-repl buffer, and a
backing inferior-lisp buffer connected to the LISP interpreter) when
invoking SBCL from within Emacs, but inferior-lisp-mode does not appear to
support Clojure and so this is necessary in order to use Clojure with
SLIME.

Previously, swank-clojure included an elisp file which allowed one to
invoke Clojure from Emacs, but this was apparently unreliable (particularly
on Windows), so the functionality was removed; instead, one now starts the
swank-clojure process externally and then connects in Emacs using the
slime-connect function.

This commit also points the remote for the clojure-mode external to its new
home on GitHub.