Don't accidentally git pull .emacs.d from externals.pl
There was a glitch in the externals script which resulted in git being
invoked in elisp directories, even when those elisp directories
weren't actual git checkouts on a given machine.  Git's behavior in
this situation was to attempt to perform a pull for the repository in
the parent directory, .emacs.d.

This commit works around this issue by checking for the existence of
the .git directory within the elisp subdir before actually pulling
when performing the update command.  Similar logic may be required for
other vcs implementations as well...