Fix c-basic-offset for reals
Instead of load the ilchymis style and overriding c-basic-offset in
the cc-mode hook, add a c-basic-offset of 4 to the definition of the
ilchymis style, overriding from the default of 2 in the gnu style.
This will ensure that we have a default basic offset of 4 when editing
C source code files.