Fix frame-closing glitch when exiting emacsclients
Fixes a glitch in the server-done-hook used when exiting an emacs
client, which was causing the last-used GUI frame to be deleted as
well after closing the client frame.  It seems that Emacs 23 will take
care of closing client frames for us, though I haven't verified this
behavior against older versions.