Fix win32 compatibility issues
Corrects issues which had crept up in my time away from Windows,
breaking compatibility with the Emacs win32 build. This commit makes
semantic an optional feature, since it is not distributed with win32
Emacs, and tests for window-system as a variable rather than as a
function (because this too fails under the win32 build of 23.3).