Set menu-bar-lines to zero for terminal frames
Leaves menu-bar-mode enabled by default in init.el, and instead adds a hook
so that menu-bar-lines is set to zero for new non-graphical frames.  This
allows us to use the MacOS menu bar in graphical frames without wasting a
line on the menu in terminal clients.

Hasn't yet been verified on X.