Remove color-theme
I'd been hanging onto the old color-theme package because, unlike the new
built-in custom themes, it supported setting themes on a frame-by-frame
basis.  However, this depended on the make-variable-frame-local function,
which has been removed in GNU Emacs 26.  As a result, I can no longer take
advantage of this feature, and color-theme provides no further advantage to
me over the built-in theme support.